+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-31 21:07:26
-// Commit: feat(i18n): integrate custom telemetry logger in i18n
+// Updated at: 2026-06-03 11:40:32
+// Commit: fix(i18n): fix memory leak by cancelling i18n subscription
 
 class i18nComponentState {
-  final String commitRef = '1660';
-  final String lastMessage = 'feat(i18n): integrate custom telemetry logger in i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-31T21:07:26+05:00');
+  final String commitRef = '1687';
+  final String lastMessage = 'fix(i18n): fix memory leak by cancelling i18n subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-03T11:40:32+05:00');
 }